+++ msgsnd_test.go
@@ -11,7 +11,7 @@ import (
 )
 
 func TestMsgsnd(t *testing.T) {
-	keyFunc := func(path string, id uint64) uint64 {
+	keyFunc := func(path string, id uint) uint {
 		key, err := ipc.Ftok(path, id)
 		if err != nil {
 			t.Fatal(err)
@@ -20,10 +20,10 @@ func TestMsgsnd(t *testing.T) {
 	}
 
 	cases := []struct {
-		key  uint64
+		key  uint
 		perm int
 	}{
-		{keyFunc("/dev/null", uint64('m')), 0600},
+		{keyFunc("/dev/null", uint('m')), 0600},
 	}
 
 	for _, tt := range cases {
@@ -34,7 +34,7 @@ func TestMsgsnd(t *testing.T) {
 		if err != nil {
 			t.Fatal(err)
 		}
-		defer func(qid uint64) {
+		defer func(qid uint) {
 			err := ipc.Msgctl(qid, ipc.IPC_RMID)
 			if err != nil {
 				t.Fatal(err)
